Automove funktioniert nun richtig (getestet)

+++ TVMC/src/mypins.h
@@ -8,14 +8,14 @@
 
 class mypins {
     private:
-        uint8_t myVal;
+        uint32_t myVal;
     public:
         mypins( );
         uint8_t setMotorDir( uint8_t dir );
         uint8_t getTestPinState( );
         void setOnboardLed( uint8_t ledState );
         void setMotorSpeed( uint8_t s );
-        uint8_t getTVstate( void );
+        uint32_t getTVstate( void );
         uint8_t getFiltMotCurrent();
         static uint8_t collisionDetected;
         static uint8_t colorchanged;
